@@ -14,12 +14,16 @@
         }
     },
     "require": {
-        "psr/simple-cache": "^1.0.0",
-        "php": ">=7.4.0"
+        "psr/simple-cache": "^2.0.0",
+        "php": "^8.0"
     },
     "require-dev": {
-        "symfony/cache": "~3.3",
-        "phpunit/phpunit": "^8.5.23",
-        "phpstan/phpstan": "^1.4.6"
+        "symfony/cache": "^5.4",
+        "phpunit/phpunit": "^10.2.3",
+        "phpstan/phpstan": "^1.10.25"
+    },
+    "scripts": {
+        "test": "@php vendor/bin/phpunit",
+        "lint": "@php vendor/bin/phpstan"
     }
 }
